# reduce number of buffers used when handling hbone traffic