I noticed that tokio::io::[copy](https://docs.rs/tokio/latest/tokio/io/fn.copy.html) uses an 8k heap-allocated buffer to copy the data. The documentation suggests that if you want to control the size of the buffer being used which I believe we do you may use tokio::io::[copy_buf](https://docs.rs/tokio/latest/tokio/io/fn.copy_buf.html) which doesn't allocate it's own buffer and instead makes use of the reader's inner buffer.