@howardjohn 2023-04-24
Can you run the benchmarks before / after? there is a readme in the benchmarks folder