@howardjohn 2023-04-24
Note they may not be accurate though (https://github.com/istio/ztunnel/issues/424)