@ilrudie 2023-04-24
Sure thing. I'll post some benchmarks results shortly.