@ilrudie 2023-04-24
**System**: Intel i7, Pop!_os Linux, rust 1.69.0
**Methodology**: Switch to desired branch, run a bench (causes a build which on my machine will kick in the cpu fan and seems to produce pretty wild bench results likely due to thermal throttling beginning), wait for cpu to cool, run actual bench & record/save result.

**Master**:
```
❯ cargo bench -- --save-baseline pr-502
    Finished bench [optimized] target(s) in 0.12s
     Running benches/throughput.rs (out/rust/release/deps/throughput-3dee05ff4e2638af)
latency/direct/1        time:   [6.8257 µs 6.8659 µs 6.9208 µs]
Found 20 outliers among 100 measurements (20.00%)
  20 (20.00%) high severe
latency/tcp/1           time:   [16.988 µs 17.332 µs 17.703 µs]
Found 8 outliers among 100 measurements (8.00%)
  7 (7.00%) high mild
  1 (1.00%) high severe
latency/hbone/1         time:   [27.454 µs 27.947 µs 28.536 µs]
Found 13 outliers among 100 measurements (13.00%)
  9 (9.00%) high mild
  4 (4.00%) high severe
latency/direct/1024     time:   [7.5578 µs 7.7450 µs 7.9860 µs]
Found 14 outliers among 100 measurements (14.00%)
  11 (11.00%) high mild
  3 (3.00%) high severe
latency/tcp/1024        time:   [17.427 µs 17.640 µs 17.901 µs]
Found 12 outliers among 100 measurements (12.00%)
  8 (8.00%) high mild
  4 (4.00%) high severe
latency/hbone/1024      time:   [43.994 ms 44.038 ms 44.098 ms]
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe

throughput/direct       time:   [874.29 µs 889.72 µs 905.78 µs]
                        thrpt:  [10.781 GiB/s 10.976 GiB/s 11.170 GiB/s]
throughput/tcp          time:   [986.67 µs 993.50 µs 1.0006 ms]
                        thrpt:  [9.7599 GiB/s 9.8295 GiB/s 9.8976 GiB/s]
throughput/hbone        time:   [19.539 ms 20.288 ms 21.140 ms]
                        thrpt:  [473.03 MiB/s 492.89 MiB/s 511.79 MiB/s]
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

Benchmarking connections/direct: Warming up for 1.0000 ms
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.9s, enable flat sampling, or reduce sample count to 60.
connections/direct      time:   [65.875 µs 66.164 µs 66.443 µs]
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) high mild
  12 (12.00%) high severe
connections/tcp         time:   [602.10 µs 743.51 µs 859.99 µs]
Benchmarking connections/hbone: Warming up for 1.0000 ms
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.1s, or reduce sample count to 90.
connections/hbone       time:   [32.563 ms 36.634 ms 40.690 ms]

metrics/write           time:   [244.96 ns 245.13 ns 245.31 ns]
Found 12 outliers among 100 measurements (12.00%)
  3 (3.00%) low severe
  4 (4.00%) high mild
  5 (5.00%) high severe
metrics/encode          time:   [999.05 ns 999.80 ns 1.0006 µs]
Found 17 outliers among 100 measurements (17.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  14 (14.00%) high severe
```

**PR Branch**:
```
❯ cargo bench -- --baseline pr-502
    Finished bench [optimized] target(s) in 0.11s
     Running benches/throughput.rs (out/rust/release/deps/throughput-3dee05ff4e2638af)
latency/direct/1        time:   [7.1821 µs 7.2939 µs 7.4113 µs]
                        change: [-38.297% -18.818% +5.0130%] (p = 0.18 > 0.05)
                        No change in performance detected.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) high mild
  13 (13.00%) high severe
latency/tcp/1           time:   [16.487 µs 16.632 µs 16.797 µs]
                        change: [-3.8413% -1.9940% -0.1619%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
latency/hbone/1         time:   [27.813 µs 29.387 µs 31.414 µs]
                        change: [-3.8150% -0.2014% +3.7578%] (p = 0.93 > 0.05)
                        No change in performance detected.
Found 16 outliers among 100 measurements (16.00%)
  3 (3.00%) low mild
  6 (6.00%) high mild
  7 (7.00%) high severe
latency/direct/1024     time:   [7.6206 µs 7.7553 µs 7.9017 µs]
                        change: [-1.5633% +0.3927% +2.3123%] (p = 0.70 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
latency/tcp/1024        time:   [17.341 µs 17.605 µs 17.908 µs]
                        change: [+2.6100% +6.3933% +10.086%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
latency/hbone/1024      time:   [43.986 ms 43.998 ms 44.010 ms]
                        change: [-0.2293% -0.0893% +0.0134%] (p = 0.17 > 0.05)
                        No change in performance detected.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low severe
  5 (5.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

throughput/direct       time:   [880.47 µs 882.99 µs 885.75 µs]
                        thrpt:  [11.025 GiB/s 11.060 GiB/s 11.091 GiB/s]
                 change:
                        time:   [-2.5275% -0.7561% +1.0369%] (p = 0.45 > 0.05)
                        thrpt:  [-1.0263% +0.7619% +2.5930%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
throughput/tcp          time:   [984.61 µs 990.51 µs 996.45 µs]
                        thrpt:  [9.8005 GiB/s 9.8592 GiB/s 9.9182 GiB/s]
                 change:
                        time:   [-1.2081% -0.3013% +0.5880%] (p = 0.55 > 0.05)
                        thrpt:  [-0.5846% +0.3022% +1.2229%]
                        No change in performance detected.
throughput/hbone        time:   [17.718 ms 18.280 ms 18.887 ms]
                        thrpt:  [529.47 MiB/s 547.05 MiB/s 564.38 MiB/s]
                 change:
                        time:   [-14.330% -9.8990% -4.9861%] (p = 0.00 < 0.05)
                        thrpt:  [+5.2478% +10.987% +16.726%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

connections/direct      time:   [65.989 µs 69.176 µs 75.218 µs]
                        change: [-26.676% -15.181% -2.7392%] (p = 0.03 < 0.05)
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  3 (3.00%) high mild
  14 (14.00%) high severe
connections/tcp         time:   [122.23 µs 123.89 µs 125.78 µs]
                        change: [-74.440% -68.411% -59.331%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
Benchmarking connections/hbone: Warming up for 1.0000 ms
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.0s, or reduce sample count to 90.
connections/hbone       time:   [28.365 ms 32.485 ms 36.565 ms]
                        change: [-25.418% -11.326% +4.3887%] (p = 0.16 > 0.05)
                        No change in performance detected.

metrics/write           time:   [243.91 ns 244.13 ns 244.37 ns]
                        change: [-4.4770% -1.9400% +0.1478%] (p = 0.11 > 0.05)
                        No change in performance detected.
Found 15 outliers among 100 measurements (15.00%)
  7 (7.00%) high mild
  8 (8.00%) high severe
metrics/encode          time:   [1.0010 µs 1.0019 µs 1.0028 µs]
                        change: [-0.2146% +0.6880% +1.9678%] (p = 0.27 > 0.05)
                        No change in performance detected.
Found 18 outliers among 100 measurements (18.00%)
  2 (2.00%) low severe
  13 (13.00%) high mild
  3 (3.00%) high severe
```